requirements.txt
acres==0.2.0
ag2==0.7.2
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
asyncer==0.0.8
attrs==24.3.0
autogen==0.7.2
autogen-agentchat==0.4.3
autogen-core==0.4.3
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blis==1.2.0
boto3==1.36.5
botocore==1.36.5
build==1.2.2.post1
cachetools==5.5.1
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.5.3
chromedriver-binary==74.0.3729.6.0
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.20.0
coloredlogs==15.0.1
confection==0.1.5
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.15
dill==0.3.9
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
duckduckgo_search==7.2.1
durationpy==0.9
etelemetry==0.3.1
executing==2.2.0
fast-depends==2.4.12
fastapi==0.115.7
filelock==3.17.0
filetype==1.2.0
fitz==0.0.1.dev2
fix-busted-json==0.0.18
flatbuffers==25.1.21
fonttools==4.55.5
free_proxy==1.1.3
frontend==0.0.3
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==0.6.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
html2text==2024.2.26
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.5.2
ipython==8.31.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kubernetes==31.0.0
langchain==0.3.15
langchain-aws==0.2.11
langchain-community==0.3.15
langchain-core==0.3.31
langchain-mistralai==0.2.4
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langcodes==3.5.0
langsmith==0.3.1
language_data==1.3.0
looseversion==1.3.0
lxml==5.3.0
m3u8==6.0.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
minify_html==0.15.0
mmh3==5.0.1
monotonic==1.6
mpire==2.10.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.17
multitasking==0.0.11
murmurhash==1.0.12
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
onnxruntime==1.20.1
openai==1.60.0
opencv-python==4.11.0.86
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pdf2image==1.17.0
peewee==3.17.8
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
playwright==1.49.1
portalocker==2.10.1
posthog==3.9.2
preshed==3.0.9
primp==0.10.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
prov==2.0.1
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyautogen==0.7.2
pyclipper==1.3.0.post6
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydot==3.0.4
pyee==12.0.0
Pygments==2.19.1
pyjsparser==2.7.1
pyotp==2.9.0
pyparsing==3.2.1
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
qdrant-client==1.13.2
rapidocr-onnxruntime==1.4.4
rdflib==6.3.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.2
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
scrapegraphai==1.37.0
seaborn==0.13.2
selenium==4.28.1
semchunk==3.0.1
sentence-transformers==3.4.0
shapely==2.0.6
shellingham==1.5.4
simpleeval==1.0.3
simplejson==3.19.3
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.2
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
textblob==0.19.0
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
traitlets==5.14.3
traits==6.4.3
transformers==4.48.1
trio==0.28.0
trio-websocket==0.11.1
tweepy==4.15.0
tweety==0.1.6
twikit==2.3.1
twython==3.9.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
undetected-playwright==0.3.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vaderSentiment==3.3.2
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
webvtt-py==0.5.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
yfinance==0.2.52
zipp==3.21.0
zstandard==0.23.0
